task1/requirements.txt
mysql-connector-python==8.0.25
flask==3.1.0